refactor: Move CommandHandlerError class into classes folder.
