+++ src/classes/commands/SubCommand.ts
@@ -1,4 +1,6 @@
-import {Command, Tag} from './Command.js';
+// Note : The SubCommand class is in the Command.ts file, see the class to know why.
+
+import {Tag} from './Command.js';
 import {SubCommandContext} from './SubCommandContext.js';
 
 export type RunSubCommandFunction = (ctx: SubCommandContext) => Promise<any>;
